feat: Layout을 컴파운드 패턴으로 수정
