+++ src/pages/home.tsx
@@ -1,73 +1,54 @@
-import CategoryList from '@/components/CategoryList';
-import Footer from '@/components/Footer';
-import Header from '@/components/Header';
-import HomeItem from '@/components/HomeItem';
-import HomeItemList from '@/components/HomeItemList';
+import CategoryList from '@/components/home/CategoryList';
+import HomeItem from '@/components/home/HomeItem';
+import HomeItemList from '@/components/home/HomeItemList';
 import Layout from '@/components/Layout';
-import { Link } from 'react-router-dom';
+import Navigation from '@/components/Navigation';
+import RegisterBtn from '@/components/register/RegisterBtn';
+import { useNavigate } from 'react-router-dom';
+import { useScrollDetection } from '@/hooks/useScrollDetection';
 
 const Home = () => {
-  const HomeFooter = (
-    <Footer>
-      <nav className="flex items-center h-full">
-        <div className="flex justify-center items-center w-[11.25rem] h-[3.75rem]">
-          <img
-            src="/home_on.svg"
-            alt="homeOn"
-            className="cursor-pointer size-7"
-          />
-        </div>
-        <div className="flex justify-center items-center w-[11.25rem] h-[3.75rem]">
-          <img
-            src="/notice_off.svg"
-            alt="noticeOff"
-            className="cursor-pointer size-7"
-          />
-        </div>
-        <div className="flex justify-center items-center w-[11.25rem] h-[3.75rem]">
-          <img
-            src="/heart_off.svg"
-            alt="heartOff"
-            className="cursor-pointer size-7"
-          />
-        </div>
-        <div className="flex justify-center items-center w-[11.25rem] h-[3.75rem]">
-          <img
-            src="/my_off.svg"
-            alt="myOff"
-            className="cursor-pointer size-7"
-          />
-        </div>
-      </nav>
-    </Footer>
-  );
+  const navigate = useNavigate();
+  const { isScrolled, elementRef } = useScrollDetection(0);
+
   return (
-    <Layout header={<Header path="/">치즈 마켓</Header>} footer={HomeFooter}>
-      <Link to="/register">등록</Link>
-      <HomeItemList name="베스트 경매">
-        <HomeItem />
-        <HomeItem />
-        <HomeItem />
-        <HomeItem />
-        <HomeItem />
-      </HomeItemList>
-      <HomeItemList name="종료 임박 경매">
-        <HomeItem />
-        <HomeItem />
-        <HomeItem />
-        <HomeItem />
-        <HomeItem />
-      </HomeItemList>
-      <HomeItemList name="카테고리">
-        <CategoryList />
-      </HomeItemList>
-      <HomeItemList name="사전 등록 경매">
-        <HomeItem />
-        <HomeItem />
-        <HomeItem />
-        <HomeItem />
-        <HomeItem />
-      </HomeItemList>
+    <Layout>
+      <Layout.Header handleBack={() => navigate('/')}>치즈 마켓</Layout.Header>
+      <Layout.Main>
+        <div
+          ref={elementRef}
+          className="relative flex flex-col w-full gap-4 overflow-y-scroll"
+        >
+          <HomeItemList name="베스트 경매">
+            <HomeItem />
+            <HomeItem />
+            <HomeItem />
+            <HomeItem />
+            <HomeItem />
+          </HomeItemList>
+          <HomeItemList name="종료 임박 경매">
+            <HomeItem />
+            <HomeItem />
+            <HomeItem />
+            <HomeItem />
+            <HomeItem />
+          </HomeItemList>
+          <HomeItemList name="카테고리">
+            <CategoryList />
+          </HomeItemList>
+          <HomeItemList name="사전 등록 경매">
+            <HomeItem />
+            <HomeItem />
+            <HomeItem />
+            <HomeItem />
+            <HomeItem />
+          </HomeItemList>
+          <RegisterBtn isScrolled={isScrolled} />
+        </div>
+      </Layout.Main>
+      <Layout.Footer>
+        <Navigation active="home" />
+      </Layout.Footer>
     </Layout>
   );
 };
